refactor: rename temp to template
