+++ src/layouts/BaseLayout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import DragWindowRegion from "@/components/DragWindowRegion";
-import NavigationMenu from "@/components/temp/NavigationMenu";
+import NavigationMenu from "@/components/template/NavigationMenu";
 
 export default function BaseLayout({
   children,
