+++ src/pages/HomePage.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import ToggleTheme from "@/components/ToggleTheme";
 import { useTranslation } from "react-i18next";
 import LangToggle from "@/components/LangToggle";
-import Footer from "@/components/temp/Footer";
-import InitialIcons from "@/components/temp/InitialIcons";
+import Footer from "@/components/template/Footer";
+import InitialIcons from "@/components/template/InitialIcons";
 
 export default function HomePage() {
   const { t } = useTranslation();
